Enable login throttling
Compared to oldCAS, throttling in newCAS has more functionality. It
supports both failure and capacity throttling and it supports using
a backend such as JDBC, Redis and Hazelcast in addition to the in-
memory solution.

Current implementation for newCAS follows the oldCAS behavior.

* Throttling on IP address and username (if username is provided)
* Throttling state are stored in memory
* Same threshhold at a maximun of five failures per second

No customization is required in newCAS. However, this comes with
a trade-off that all authentication failures are throttled. Some
failures in newCAS such as two-factor / TOS consent are used for
state transition only. However, this is not a problem for newCAS
with the five failures per second setting since both entering the
one-time code or checking the consent check box by a human should
not trigger the throttle at all.